Clean code % Fix Phi3 bug

+++ internvl_chat/eval/caption/evaluate_caption.py
@@ -154,7 +154,7 @@ def evaluate_chat_model():
 
         image_ids, captions = [], []
         for _, (pixel_values, ids, _, _) in tqdm(enumerate(dataloader)):
-            pixel_values = pixel_values.to(torch.bfloat16).cuda()
+            pixel_values = pixel_values.to(torch.float16).cuda()
             generation_config = dict(
                 num_beams=args.num_beams,
                 max_new_tokens=ds_collections[ds_name]['max_new_tokens'],
@@ -254,7 +254,7 @@ if __name__ == '__main__':
 
     tokenizer = AutoTokenizer.from_pretrained(args.checkpoint, trust_remote_code=True, use_fast=False)
     model = InternVLChatModel.from_pretrained(
-        args.checkpoint, low_cpu_mem_usage=True, torch_dtype=torch.bfloat16,
+        args.checkpoint, low_cpu_mem_usage=True, torch_dtype=torch.float16,
         load_in_8bit=args.load_in_8bit).eval()
     if not args.load_in_8bit:
         model = model.cuda()
